requirements.txt
perceval
python-dotenv
openai
langchain[openai]
google-generativeai
ollama
jupyter
pandas
matplotlib
seaborn

